# odin-project-recipes odin project data!
hello this is me vaibhav and im trying all the git commands by my self.

what have i learned so far is the we have to first clone the reository we want to work on like in my case i want to work on odin-project-recipes.
: git clone "code of ssh"
then the directory is downloaded to the present working directory.
the we made changes to the directory or create a new file or delete some.

after making changes to the desired files we will first do is:

:git